create routes folder
